+++ src/hmtrp/hmtrp.h
@@ -124,7 +124,7 @@ namespace upm {
      * waiting forever (default).
      * @return Number of bytes read; 0 if timed out and millis is >= 0
      */
-    int readData(char *buffer, size_t len, int millis=-1);
+    int readData(char *buffer, int len, int millis=-1);
 
     /**
      * Writes the data in the buffer to the device
@@ -133,7 +133,7 @@ namespace upm {
      * @param len Length of the buffer
      * @return Number of bytes written
      */
-    int writeData(char *buffer, size_t len);
+    int writeData(char *buffer, int len);
 
     /**
      * Sets up proper tty I/O modes and the baud rate. The default
